+++ test/rounded_loading_button_test.dart
@@ -5,21 +5,99 @@ import 'package:rounded_loading_button/rounded_loading_button.dart';
 
 void main() {
 
-  testWidgets('should show progress indicator when loading', (tester) async {
-    final btnController = RoundedLoadingButtonController();
+  testWidgets('should show progress indicator when in loading state', (tester) async {
+    final btnController = new RoundedLoadingButtonController();
     
     await tester.pumpWidget(
-      RoundedLoadingButton(
-        child: Text('Tap me!', style: TextStyle(color: Colors.white)),
-        controller: btnController,
-        width: 200,
-        color: Colors.purple,
+      Directionality(
+        textDirection: TextDirection.ltr,
+        child: Material(
+          child: Center(
+            child: RoundedLoadingButton(
+              child: Text('Tap me!', style: TextStyle(color: Colors.white)),
+              controller: btnController,
+              width: 200
+            ),
+          ),
+        )
       )
     );
 
     btnController.start();
-    
-    expect(find.byType(Text), findsOneWidget);
+
+    await tester.pump(const Duration(seconds: 1));
+
     expect(find.byType(CircularProgressIndicator), findsOneWidget);
+  });
+
+  testWidgets('should not show progress indicator when in idle state', (tester) async {
+    final btnController = new RoundedLoadingButtonController();
+    
+    await tester.pumpWidget(
+      Directionality(
+        textDirection: TextDirection.ltr,
+        child: Material(
+          child: Center(
+            child: RoundedLoadingButton(
+              child: Text('Tap me!', style: TextStyle(color: Colors.white)),
+              controller: btnController,
+              width: 200
+            ),
+          ),
+        )
+      )
+    );
+
+    expect(find.byType(CircularProgressIndicator), findsNothing);
+  });
+
+  testWidgets('should show icon when in success state', (tester) async {
+    final btnController = new RoundedLoadingButtonController();
+    
+    await tester.pumpWidget(
+      Directionality(
+        textDirection: TextDirection.ltr,
+        child: Material(
+          child: Center(
+            child: RoundedLoadingButton(
+              child: Text('Tap me!', style: TextStyle(color: Colors.white)),
+              controller: btnController,
+              width: 200
+            ),
+          ),
+        )
+      )
+    );
+
+    btnController.success();
+
+    await tester.pumpAndSettle();
+
+    expect(find.byIcon(Icons.check), findsOneWidget);
+  });
+
+  testWidgets('should show icon when in error state', (tester) async {
+    final btnController = new RoundedLoadingButtonController();
+    
+    await tester.pumpWidget(
+      Directionality(
+        textDirection: TextDirection.ltr,
+        child: Material(
+          child: Center(
+            child: RoundedLoadingButton(
+              child: Text('Tap me!', style: TextStyle(color: Colors.white)),
+              controller: btnController,
+              width: 200
+            ),
+          ),
+        )
+      )
+    );
+
+    btnController.error();
+
+    await tester.pumpAndSettle();
+
+    expect(find.byIcon(Icons.close), findsOneWidget);
   });
 }